+++ core/src/main/java/ru/tinkoff/acquiring/sdk/requests/AcquiringRequest.kt
@@ -20,10 +20,13 @@ import com.google.gson.Gson
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.CompletableDeferred
+import kotlinx.coroutines.Deferred
 import okhttp3.Response
 import ru.tinkoff.acquiring.sdk.AcquiringSdk
 import ru.tinkoff.acquiring.sdk.exceptions.NetworkException
 import ru.tinkoff.acquiring.sdk.network.AcquiringApi
+import ru.tinkoff.acquiring.sdk.network.AcquiringApi.JSON
 import ru.tinkoff.acquiring.sdk.network.NetworkClient
 import ru.tinkoff.acquiring.sdk.responses.AcquiringResponse
 import ru.tinkoff.acquiring.sdk.utils.Request
@@ -37,7 +40,8 @@ import java.security.PublicKey
  *
  * @author Mariya Chernyadieva, Taras Nagorny
  */
-abstract class AcquiringRequest<R : AcquiringResponse>(internal val apiMethod: String) : Request<R> {
+abstract class AcquiringRequest<R : AcquiringResponse>(internal val apiMethod: String) :
+    Request<R> {
 
     protected val gson: Gson = NetworkClient.createGson()
 
@@ -46,13 +50,14 @@ abstract class AcquiringRequest<R : AcquiringResponse>(internal val apiMethod: S
 
     internal lateinit var terminalKey: String
     internal lateinit var publicKey: PublicKey
+
     @Volatile
     private var disposed = false
     private val ignoredFieldsSet: HashSet<String> = hashSetOf(DATA, RECEIPT, RECEIPTS, SHOPS)
+    private val headersMap: HashMap<String, String> = hashMapOf()
 
     internal open val tokenIgnoreFields: HashSet<String>
         get() = ignoredFieldsSet
-
 
     protected abstract fun validate()
 
@@ -72,13 +77,37 @@ abstract class AcquiringRequest<R : AcquiringResponse>(internal val apiMethod: S
         return map
     }
 
-    protected fun <R : AcquiringResponse> performRequest(request: AcquiringRequest<R>,
-                                                         responseClass: Class<R>,
-                                                         onSuccess: (R) -> Unit,
-                                                         onFailure: (Exception) -> Unit) {
+    protected fun <R : AcquiringResponse> performRequest(
+        request: AcquiringRequest<R>,
+        responseClass: Class<R>,
+        onSuccess: (R) -> Unit,
+        onFailure: (Exception) -> Unit
+    ) {
         request.validate()
         val client = NetworkClient()
         client.call(request, responseClass, onSuccess, onFailure)
+    }
+
+    open fun performRequestAsync(responseClass: Class<R>): Deferred<Result<R>> {
+        this.validate()
+        val client = NetworkClient()
+        val deferred: CompletableDeferred<Result<R>> = CompletableDeferred()
+
+        client.call(this, responseClass,
+            onSuccess = {
+                deferred.complete(Result.success(it))
+            },
+            onFailure = {
+                deferred.complete(Result.failure(it))
+            })
+        return deferred
+    }
+
+    suspend fun performSuspendRequest(responseClass: Class<R>): Result<R> {
+        return performRequestAsync(responseClass).run {
+            start()
+            await()
+        }
     }
 
     protected fun <R : AcquiringResponse> performRequestFlow(request: AcquiringRequest<R>,
@@ -129,8 +158,18 @@ abstract class AcquiringRequest<R : AcquiringResponse>(internal val apiMethod: S
         }
     }
 
+    fun addUserAgentHeader(userAgent: String = System.getProperty("http.agent")) {
+        headersMap.put("User-Agent", userAgent)
+    }
+
+    fun addContentHeader(content: String = JSON) {
+        headersMap.put("Accept", content)
+    }
+
     protected open fun getToken(): String? =
         AcquiringSdk.tokenGenerator?.generateToken(this, paramsForToken())
+
+    internal fun getHeaders() = headersMap
 
     private fun paramsForToken(): MutableMap<String, Any> {
         val tokenParams = asMap()
@@ -190,7 +229,7 @@ abstract class AcquiringRequest<R : AcquiringResponse>(internal val apiMethod: S
         const val REQUEST_KEY = "RequestKey"
         const val SOURCE = "Source"
         const val PAYMENT_SOURCE = "PaymentSource"
-        const val ANDROID_PAY_TOKEN = "EncryptedPaymentData"
+        const val ENCRYPTED_PAYMENT_DATA = "EncryptedPaymentData"
         const val DATA_TYPE = "DataType"
         const val REDIRECT_DUE_DATE = "RedirectDueDate"
         const val NOTIFICATION_URL = "NotificationURL"
@@ -204,5 +243,10 @@ abstract class AcquiringRequest<R : AcquiringResponse>(internal val apiMethod: S
         const val THREE_DS_SERVER_TRANS_ID = "threeDSServerTransID"
         const val TRANS_STATUS = "transStatus"
         const val CRES = "cres"
+        const val PAYSOURCE = "Paysource"
     }
+}
+
+suspend inline fun <reified R : AcquiringResponse> AcquiringRequest<R>.performSuspendRequest(): Result<R> {
+    return performSuspendRequest(R::class.java)
 }