+++ ui/src/main/java/ru/tinkoff/acquiring/sdk/TinkoffAcquiring.kt
@@ -23,31 +23,27 @@ import android.content.Intent
 import androidx.activity.result.contract.ActivityResultContract
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
+import kotlinx.coroutines.*
 import ru.tinkoff.acquiring.sdk.localization.LocalizationSource
-import ru.tinkoff.acquiring.sdk.models.AsdkState
-import ru.tinkoff.acquiring.sdk.models.DefaultState
-import ru.tinkoff.acquiring.sdk.models.FpsState
-import ru.tinkoff.acquiring.sdk.models.GooglePayParams
-import ru.tinkoff.acquiring.sdk.models.PaymentSource
+import ru.tinkoff.acquiring.sdk.models.*
 import ru.tinkoff.acquiring.sdk.models.options.FeaturesOptions
-import ru.tinkoff.acquiring.sdk.models.options.screen.AttachCardOptions
-import ru.tinkoff.acquiring.sdk.models.options.screen.BaseAcquiringOptions
-import ru.tinkoff.acquiring.sdk.models.options.screen.PaymentOptions
-import ru.tinkoff.acquiring.sdk.models.options.screen.SavedCardsOptions
+import ru.tinkoff.acquiring.sdk.models.options.screen.*
 import ru.tinkoff.acquiring.sdk.models.paysources.AttachedCard
 import ru.tinkoff.acquiring.sdk.models.paysources.CardData
 import ru.tinkoff.acquiring.sdk.models.paysources.GooglePay
 import ru.tinkoff.acquiring.sdk.payment.PaymentProcess
+import ru.tinkoff.acquiring.sdk.requests.performSuspendRequest
+import ru.tinkoff.acquiring.sdk.responses.GetTerminalPayMethodsResponse
+import ru.tinkoff.acquiring.sdk.responses.TerminalInfo
 import ru.tinkoff.acquiring.sdk.redesign.cards.list.ui.CardsListActivity
 import ru.tinkoff.acquiring.sdk.responses.TinkoffPayStatusResponse
+import ru.tinkoff.acquiring.sdk.ui.activities.*
 import ru.tinkoff.acquiring.sdk.ui.activities.AttachCardActivity
 import ru.tinkoff.acquiring.sdk.ui.activities.BaseAcquiringActivity
 import ru.tinkoff.acquiring.sdk.ui.activities.NotificationPaymentActivity
 import ru.tinkoff.acquiring.sdk.ui.activities.PaymentActivity
 import ru.tinkoff.acquiring.sdk.ui.activities.QrCodeActivity
+import ru.tinkoff.acquiring.sdk.ui.activities.YandexPaymentActivity
 
 /**
  * Точка входа для взаимодействия с Acquiring SDK
@@ -144,6 +140,23 @@ class TinkoffAcquiring(
     /**
      * Запуск экрана Acquiring SDK для проведения оплаты
      *
+     * @param activity        контекст для запуска экрана из Activity
+     * @param options         настройки платежной сессии и визуального отображения экрана
+     * @param requestCode     код для получения результата, по завершению работы экрана Acquiring SDK
+     * @param yandexPayToken  параметр платежной сессии от яндекса
+     */
+    fun openYandexPaymentScreen(activity: Activity,
+                          options: PaymentOptions,
+                          requestCode: Int,
+                          yandexPayToken: String) {
+        options.asdkState = YandexPayState(yandexPayToken)
+        val intent = prepareIntent(activity, options, YandexPaymentActivity::class.java)
+        activity.startActivityForResult(intent, requestCode)
+    }
+
+    /**
+     * Запуск экрана Acquiring SDK для проведения оплаты
+     *
      * @param fragment    контекст для запуска экрана из Fragment
      * @param options     настройки платежной сессии и визуального отображения экрана
      * @param requestCode код для получения результата, по завершению работы экрана Acquiring SDK
@@ -197,11 +210,30 @@ class TinkoffAcquiring(
         onFailure: ((Throwable) -> Unit)? = null
     ) {
         CoroutineScope(Dispatchers.IO).launch {
-            sdk.tinkoffPayStatus().execute({
-                launch(Dispatchers.Main) { onSuccess(it) }
-            }, {
-                launch(Dispatchers.Main) { onFailure?.invoke(it) }
-            })
+            val mainScope = this
+            val result = sdk.tinkoffPayStatus().performSuspendRequest()
+            withContext(Dispatchers.Main) {
+                result.fold(onSuccess = onSuccess, onFailure = { onFailure?.invoke(it) })
+                mainScope.cancel()
+            }
+        }
+    }
+
+    /**
+     * Проверка доступных спосбов оплаты
+     */
+    fun checkTerminalInfo(onSuccess: (TerminalInfo?) -> Unit,
+                          onFailure: ((Throwable) -> Unit)? = null
+    ) {
+        CoroutineScope(Dispatchers.IO).launch {
+            val mainScope = this
+            val result = sdk.getTerminalPayMethods()
+                .performSuspendRequest()
+                .map { it.terminalInfo }
+            withContext(Dispatchers.Main) {
+                result.fold(onSuccess = onSuccess, onFailure = { onFailure?.invoke(it) })
+                mainScope.cancel()
+            }
         }
     }