+++ ui/src/main/java/ru/tinkoff/acquiring/sdk/models/options/screen/AttachCardOptions.kt
@@ -56,3 +56,4 @@ class AttachCardOptions : BaseCardsOptions<AttachCardOptions>, Parcelable {
         }
     }
 }
+